restructure and objects endpoints updated
