+++ weaviate-community/src/collections/mod.rs
@@ -1,0 +1,2 @@
+pub mod schema;
+pub mod objects;
