+++ weaviate-community/src/schema.rs
@@ -1,4 +1,4 @@
-use crate::collections::{Class, Property, ShardStatus, Tenant};
+use crate::collections::schema::{Class, Property, ShardStatus, Tenant};
 use reqwest::{Response, Url};
 use std::error::Error;
 
@@ -68,7 +68,7 @@ impl Schema {
     /// POST /v1/schema
     /// ```
     /// use weaviate_community::Client;
-    /// use weaviate_community::collections::Class;
+    /// use weaviate_community::collections::schema::Class;
     ///
     /// #[tokio::main]
     /// async fn main() {
@@ -181,7 +181,7 @@ impl Schema {
         endpoint.push_str(shard_name);
         let endpoint = self.endpoint.join(&endpoint)?;
         let payload = serde_json::json!({
-                "status": status.value()
+            "status": status
         });
         let res = self.client.put(endpoint).json(&payload).send().await?;
         Ok(res)
@@ -259,7 +259,7 @@ mod tests {
     // Tests currently require a weaviate instance to be running on localhost, as I have not yet
     // implemented anything to mock the database. In future, actual tests will run as integration
     // tests in a container as part of the CICD process.
-    use crate::collections::{
+    use crate::collections::schema::{
         ActivityStatus, Class, MultiTenancyConfig, Property, ShardStatus, Tenant,
     };
     use crate::Client;
